+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 0.3.0.0
+**Breaking API Changes**
+In this version, TermPair clients from previous versions cannot connect to this TermPair server
+
 * Use new key sharing scheme: Different keys used in different directions; keys rotated; secret keys retrieved via RSA public key rather than embedding in the URL
 * [bugfix] Terminal dimensions in browser match upon initial connection, instead of after resizing
+* Allow static site to route terminal traffic through other server. If static site is detected, user can enter the terminal id and server url in the browser UI.
+* Allow Terminal ID to be entered on landing page
+
 
 ## 0.2.0.0
 * Add ability to copy+paste using keystrokes (copy with ctrl+shift+c or ctrl+shift+x, and paste with ctrl+shift+v)